cv/requirements.txt
matplotlib==3.2.2
mayavi==4.8.1
numpy==1.20.1
Pillow==9.5.0
scipy==1.9.3

